+++ CHANGELOG.md
@@ -15,9 +15,9 @@ our [guidelines for writing a good changelog entry](https://github.com/biomejs/b
 
 #### New features
 
-- Allow passing values to suppression comments. This is used for the lint rule
-  `useExhaustiveDependencies`, which is now able to suppress specific
-  dependencies. Fixes #2509. Contributed by @arendjr
+- Allow suppression comments to suppress individual instances of rules. This is
+  used for the lint rule `useExhaustiveDependencies`, which is now able to
+  suppress specific dependencies. Fixes #2509. Contributed by @arendjr
 
 #### Enhancements